# Conflicts:
#	Program.cs
#	SQL_Query/populatedb.sql

+++ Core/Entities/Order.cs
@@ -5,9 +5,8 @@ namespace Database_project.Core.Entities;
 public class Order
 {
     [Key]
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
     public string AirlineConfirmationNumber { get; set; }
 
     public ICollection<Ticket> Tickets { get; set; }
-
 }
